mm: memcontrol: per-lruvec stats infrastructure fix 2
Even with the previous fix routing !page->mem_cgroup stats to the root
cgroup, we still see crashes in certain configurations as the root is not
initialized for the earliest possible accounting sites in certain
configurations.

Don't track uncharged pages at all, not even in the root.  This takes care
of early accounting as well as special pages that aren't tracked.

Because we still need to account at the pgdat level, we can no longer
implement the lruvec_page_state functions on top of the lruvec_state ones.
But that's okay.  It was a little silly to look up the nodeinfo and
descend to the lruvec, only to container_of() back to the nodeinfo where
the lruvec_stat structure is sitting.

Link: http://lkml.kernel.org/r/20170605175254.GA8547@cmpxchg.org
Signed-off-by: Johannes Weiner <hannes@cmpxchg.org>
Reported-by: Guenter Roeck <linux@roeck-us.net>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Vladimir Davydov <vdavydov.dev@gmail.com>
Cc: Rik van Riel <riel@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>