mm, page_owner: don't grab zone->lock for init_pages_in_zone()
init_pages_in_zone() is run under zone->lock, which means a long lock time
and disabled interrupts on large machines.  This is currently not an issue
since it runs early in boot, but a later patch will change that.  However,
like other pfn scanners, we don't actually need zone->lock even when other
cpus are running.  The only potentially dangerous operation here is
reading bogus buddy page owner due to race, and we already know how to
handle that.  The worse that can happen is that we skip some early
allocated pages, which should not affect the debugging power of page_owner
noticeably.

Link: http://lkml.kernel.org/r/20170720134029.25268-4-vbabka@suse.cz
Signed-off-by: Vlastimil Babka <vbabka@suse.cz>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: Michal Hocko <mhocko@kernel.org>
Cc: Mel Gorman <mgorman@techsingularity.net>
Cc: Yang Shi <yang.shi@linaro.org>
Cc: Laura Abbott <labbott@redhat.com>
Cc: Vinayak Menon <vinmenon@codeaurora.org>
Cc: zhong jiang <zhongjiang@huawei.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>